+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+* Added `--add-columns` flag to the `fill_table` management command. It creates all the
+  field types before filling the table with random data.
+
 ## Released (2021-04-08)
 
 * Added support for importing tables from XML files.